+++ HISTORY.md
@@ -1,8 +1,9 @@
 # History
 
 ## 0.7.0 (2021-TBD) 
-* Add a filesystem cache
+* Add a filesystem backend
 * Add option to manually cache response objects
+* Add `use_temp` option to both SQLite and filesystem backends to store files in a temp directory
 * Use thread-local connections for SQLite backend
 * Fix `DynamoDbDict.__iter__` to return keys instead of values